% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNames.R
\name{checkNames}
\alias{checkNames}
\title{Internal function that check the names.}
\usage{
checkNames(df = NULL, baselines = NULL)
}
\arguments{
\item{df}{named list with the variables to be checked for.}

\item{baselines}{integer stating the number of baselines.}
}
\description{
Not intended to be used by the user.
}
\examples{

}

