% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credibilityIntervals.R
\name{credibilityIntervals}
\alias{credibilityIntervals}
\title{Title}
\usage{
credibilityIntervals(df, df2 = NULL, xlab = "Bayesian models",
  ylab1 = "Posterior Trophic Position", ylab2 = "Posterior alpha")
}
\arguments{
\item{df}{}
}
\description{
Title
}
