% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyTP.isotopeData.R
\name{easyTP}
\alias{easyTP}
\title{Title}
\usage{
easyTP(siData = siData, lambda = 2, print = TRUE, n.adapt = 10000,
  n.iter = 10000, n.chains = 2, models = c("oneBaseline", "twoBaselines",
  "twoBaselinesFull"))
}
\arguments{
\item{print}{}
}
\description{
Title
}

