% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIsotopeData.R
\name{extractIsotopeData}
\alias{extractIsotopeData}
\title{Title}
\usage{
extractIsotopeData(df = NULL, b1 = "Baseline 1", b2 = NULL,
  baselineColumn = "FG", speciesColumn = "Spp", communityColumn = NULL,
  deltaC = NULL, deltaN = NULL, d13C = "d13C", d15N = "d15N",
  seed = 666)
}
\arguments{
\item{deltaN}{}
}
\description{
Title
}
