% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finnish_Lakes.R
\docType{data}
\name{Finnish_Lakes}
\alias{Finnish_Lakes}
\title{Data frame of food webs in Inari and Kilpis Lakes (Finland)}
\format{A data frame with 7 variables:
\describe{
  \item{Lake}{factor, with two levels, each representing one Lake}
  \item{Species.group}{factor, with 20 levels, each representing one species}
  \item{d13C}{numeric, representing delta 13 C isotope values}
  \item{d15N}{numeric, representing delta 15 N isotope values}
  \item{C}{numeric, ammount of carbon}
  \item{N}{numeric, ammount of nitrogen}
  \item{C.N}{numeric, C/N ratio}
}}
\usage{
data("Finnish_Lakes")
}
\description{
Dataset of stable isotope values from two relatively large and deep
oligotrophic Finnish lakes Inari and Kilpis.
}
\keyword{datasets}
