% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generataIsotopeData.R
\name{generateIsotopeData}
\alias{generateIsotopeData}
\title{A function to generate stable isotope data for analysis in R}
\usage{
generateIsotopeData(matrix = NULL, labels = NULL, iso_labels = NULL,
  seed = 3, ...)
}
\value{
Missing.
}
\description{
This function generates random stable isotope data (unlimited elements) to use
within the stable isotope R packages.
}
\examples{

}
