% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMode.R
\name{getPosteriorMode}
\alias{getPosteriorMode}
\title{Function to get mode from posterior distribution}
\usage{
getPosteriorMode(df = NULL)
}
\arguments{
\item{df}{data.frame with posterior distributions}
}
\value{
a list with modes

#@importFrom magrittr "%>%"
}
\description{
This function is a wrapper of \code{\link[hdrcde]{hdr}}, it returns one
mode (if receives a vector), otherwise returns a lists of modes
}
\examples{
# List example
a <- list("First" = rnorm(100,1), "Second" = rnorm(100,2))
getPosteriorMode(a)

# vector example
getPosteriorMode(rnorm(100,5))
}
