% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopeData.R
\name{isotopeData}
\alias{isotopeData}
\title{Construct isotope data from a data frame}
\usage{
isotopeData(df, d15N = "d15N", d13C = "d13C", grouping = "FG",
  consumer = NULL, b1 = NULL, b2 = NULL, TEF = NULL)
}
\arguments{
\item{TEF}{}
}
\description{
Construct isotope data from a data frame
}
