% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsBayesianModel.R
\name{jagsBayesianModel}
\alias{jagsBayesianModel}
\title{Return a JAGS Bayesian model to use with tRophicPosition.}
\usage{
jagsBayesianModel(model = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Description
}
