% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsOneBaseline.R
\name{jagsOneBaseline}
\alias{jagsOneBaseline}
\title{Defines the jags model to fit the single baseline trohpic position model}
\usage{
jagsOneBaseline(muBprior = NULL, sigmaBprior = NULL, muDeltaN = NULL,
  sigmaDeltaN = NULL, sigma = NULL, TP = NULL, lambda = NULL)
}
\arguments{
\item{TP}{}

\item{lambda}{an integer indicating the trophic position of the baseline.}
}
\value{
A jags model as a character string
}
\description{
Takes some parameters and returns a jags model object as a
character string for passing to \code{\link[rjags.model]{rjags.model}}.
}

