% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsTPModels.R
\name{jagsTPmodels}
\alias{jagsTPmodels}
\title{Return a JAGS model with one to three baselines.}
\usage{
jagsTPmodels(nb = 1, ...)
}
\arguments{
\item{nb}{integer, numbers of baselines required.}
}
\description{
To do.
}

