% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsTwoBaselines.R
\name{jagsTwoBaselines}
\alias{jagsTwoBaselines}
\title{Defines the jags model to fit the two baseline trophic position model}
\usage{
jagsTwoBaselines(TPprior = NULL, mubprior = NULL)
}
\arguments{
\item{mubprior}{}
}
\value{
A jags model as a character string
}
\description{
Takes some parameters and returns a jags model object as a
character string for passing to \code{\link[rjags.model]{rjags.model}}.
}

