% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadIsotopeData.R
\name{loadIsotopeData}
\alias{loadIsotopeData}
\title{Extract and load stable isotope data for one consumer from a data frame}
\usage{
loadIsotopeData(df = NULL, consumer = NULL, b1 = "Baseline 1",
  b2 = NULL, groupingColumn = "FG", d13C = "d13C", d15N = "d15N",
  deltaC = NULL, deltaN = NULL, seed = 666)
}
\arguments{
\item{seed}{}
}
\description{
Extract and load stable isotope data for one consumer from a data frame
}
