% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orestias.R
\docType{data}
\name{Orestias}
\alias{Orestias}
\title{Named list containing stable isotope values of Orestias chungarensis}
\format{
A named list with 8 elements:
\describe{
  \item{dCc}{numeric, stable isotope values of d13C for Orestias
  chungarensis}
  \item{dNc}{numeric, stable isotope values of d15N for Orestias
  chungarensis}
  \item{dCb1}{numeric, stable isotope values of d13C for baseline 1}
  \item{dNb1}{numeric, stable isotope values of d15N for baseline 1}
  \item{dCb2}{numeric, stable isotope values of d13C for baseline 2}
  \item{dNb2}{numeric, stable isotope values of d15N for baseline 2}
  \item{deltaN}{numeric, trophic discrimination factor for nitrogen}
  \item{deltaC}{numeric, trophic discrimination factor for carbon}
}
}
\usage{
data("Orestias")
}
\description{
A dataset containing stable isotope values (d13C and d15N) for the killifish
*Orestias chungarensis*
(http://www.fishbase.se/summary/Orestias-chungarensis.html), an endemic fish
found in the Chungara Lake at 4,500 meters above sea level, Chilean
Altiplano. Harrod, C. & I. Vila unpublished results.
}
\keyword{datasets}
