% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.isotopeData.R
\name{plot.isotopeData}
\alias{plot.isotopeData}
\title{Plot stable isotope data (2 elements) with one or two baselines}
\usage{
\method{plot}{isotopeData}(isotopeData, consumer = NULL, b1 = NULL,
  b2 = NULL, legend = c(1.15, 1.15), density = "both")
}
\arguments{
\item{density}{}
}
\description{
Plot stable isotope data (2 elements) with one or two baselines
}
\examples{

}
