% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMCMC.R
\name{plotMCMC}
\alias{plotMCMC}
\title{Internal function that plot a mcmc.list object.}
\usage{
plotMCMC(
  x,
  trace = TRUE,
  density = TRUE,
  smooth = TRUE,
  bwf,
  auto.layout = TRUE,
  ask = graphics::par("ask"),
  ...
)
}
\arguments{
\item{x}{null}

\item{trace}{null}

\item{density}{null}

\item{smooth}{null}

\item{bwf}{null}

\item{auto.layout}{null}

\item{ask}{null}

\item{...}{null}
}
\description{
Not intended to be used by the user.
}
