% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTP.R
\name{plotTP}
\alias{plotTP}
\title{Function to plot a trophic position distribution}
\usage{
plotTP(TPdist = NULL)
}
\arguments{
\item{TPdist}{}
}
\description{
Function to plot a trophic position distribution
}
\examples{

}

