% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorTP.R
\name{posteriorTP}
\alias{posteriorTP}
\title{Function to generate posterior samples of trophic position}
\usage{
posteriorTP(model, variable.names = c("TP", "muDeltaN"), n.iter = 10000)
}
\arguments{
\item{n.iter}{}
}
\description{
Function to generate posterior samples of trophic position
}

