% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reOrder.R
\name{reOrder}
\alias{reOrder}
\title{Internal function that re-order variables from a data frame.}
\usage{
reOrder(isotopeData = NULL)
}
\arguments{
\item{isotopeData}{data frame containing the data to be ordered.}
}
\description{
Not intended to be used by the user.
}
