% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenIsotopeData.R
\name{screenIsotopeData}
\alias{screenIsotopeData}
\title{Function to plot and screen Isotope Data with one or two baselines}
\usage{
screenIsotopeData(iso1 = NULL, iso2 = NULL, b1 = NULL, b2 = NULL)
}
\arguments{
\item{b2}{}
}
\description{
Function to plot and screen Isotope Data with one or two baselines
}

