% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenIsotopeData1source.R
\name{screenIsotopeData1source}
\alias{screenIsotopeData1source}
\title{Title}
\usage{
screenIsotopeData1source(isotopeData = NULL, histogram = T)
}
\arguments{
\item{isotopeData}{}
}
\description{
Title
}
\examples{

}

