% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenIsotopeData2sources.R
\name{screenIsotopeData2sources}
\alias{screenIsotopeData2sources}
\title{Title}
\usage{
screenIsotopeData2sources(isotopeData = NULL, density = "both")
}
\arguments{
\item{density}{}
}
\description{
Title
}
\note{

}
\examples{

}

