% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenIsotopeData3sources.R
\name{screenIsotopeData3sources}
\alias{screenIsotopeData3sources}
\title{Internal function that plots the data}
\usage{
screenIsotopeData3sources(isotopeData = NULL, density = "both")
}
\arguments{
\item{isotopeData}{a named list composed at least of 6 vectors, dNb1, dCb1,
dCc, dNc, deltaN and deltaC.}

\item{density}{a character that might plot the density function.}
}
\description{
This function is intended to be used within the package. For details refer to
\code{\link{screenIsotopeData}}. This function plots 3 baselines coupled with
the siar version of tRophicPosition.
}
\note{

}
\examples{

}
