% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenIsotopeDataMoreSources.R
\name{screenIsotopeDataMoreSources}
\alias{screenIsotopeDataMoreSources}
\title{Internal function that plots the data}
\usage{
screenIsotopeDataMoreSources(isotopeData = NULL, density = "both",
  baselines = 2, consumer = consumer, b1 = b1, b2 = b2,
  legend = legend, title = NULL, ...)
}
\arguments{
\item{isotopeData}{an isotopeData class object.}

\item{density}{string. Can be "both", "above", "right" or "none".}

\item{baselines}{number of baselines plotted.}

\item{consumer}{string representing consumer}

\item{b1}{string representing baseline 1}

\item{b2}{string representing baseline 2}

\item{legend}{string representing legend}

\item{title}{string representing title}

\item{...}{additional arguments passed to this function}
}
\description{
This function is intended to be used within the package. For details refer to
\code{\link{screenIsotopeData}}.
}
