% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMCMC.R
\name{set.mfrow}
\alias{set.mfrow}
\title{Internal function that set.mfrow.}
\usage{
set.mfrow(Nchains = 1, Nparms = 1, nplots = 1, sepplot = FALSE)
}
\arguments{
\item{Nchains}{null}

\item{Nparms}{null}

\item{nplots}{null}

\item{sepplot}{null}
}
\description{
Not intended to be used by the user.
}
