% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanOneBaseline.R
\name{stanOneBaseline}
\alias{stanOneBaseline}
\title{Title}
\usage{
stanOneBaseline(
  d15N.bsl = NULL,
  d15N.con = NULL,
  Spp = NULL,
  LANCE = NULL,
  TDF = 3.4,
  TDF_sd = 0.51,
  lambda = 1,
  iter = 10000
)
}
\arguments{
\item{lambda}{}
}
\description{
Title
}
