% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanOneBaselineCSIAA.R
\name{stanOneBaselineCSIA}
\alias{stanOneBaselineCSIA}
\title{Title}
\usage{
stanOneBaselineCSIA(aminoacid = NULL, Spp = NULL, TDF = NULL, TDF_sd = 1.2)
}
\arguments{
\item{aminoacid}{aminoacid}

\item{TDF}{TDF}

\item{TDF_sd}{TDF_sd}
}
\value{
value
}
\description{
Title
}
