% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.isotopeData.R
\name{summary}
\alias{summary}
\title{Summary for stable isotope data}
\usage{
summary(siData, print = TRUE, round_dec = 1)
}
\arguments{
\item{siData}{an isotopeData class object.}

\item{print}{a logical value to indicate whether the summary is printed.}

\item{round_dec}{number of decimals keeped.}
}
\value{
a list with number of observations, mean, standard deviation,
standard error, minimum, maximum and median for each element of an
isotopeData class object.
}
\description{
Summary for stable isotope data
}
\examples{
## S3 method for class 'isotopeData'
a <- generateTPData()
summary(a)

}
