% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toStacked.R
\name{toStacked}
\alias{toStacked}
\title{Internal function that stack variables from a named list into a data frame.}
\usage{
toStacked(isotopeData = NULL, baselines = 1)
}
\arguments{
\item{isotopeData}{a named list composed at least of 6 vectors, dNb1, dCb1,
dCsc, dNsc, deltaN and deltaC}

\item{baselines}{integer stating the number of baselines.}
}
\description{
Not intended to be used by the user.
}

