% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPmodel.R
\name{TPmodel}
\alias{TPmodel}
\title{Function to build a JAGS-based Bayesian model to calculate trophic position}
\usage{
TPmodel(data = NULL, model.string = NULL, n.chains = 2, n.adapt = 10000)
}
\arguments{
\item{n.adapt}{}
}
\description{
Function to build a JAGS-based Bayesian model to calculate trophic position
}

