% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trout.R
\docType{data}
\name{Trout}
\alias{Trout}
\title{Named list containing stable isotope values of Oncorhynchus mykiss}
\format{A named list with 8 elements:
\describe{
  \item{dCc}{numeric, stable isotope values of d13C for Oncorhynchus mykiss}
  \item{dNc}{numeric, stable isotope values of d15N for Oncorhynchus mykiss}
  \item{dCb1}{numeric, stable isotope values of d13C for baseline 1}
  \item{dNb1}{numeric, stable isotope values of d15N for baseline 1}
  \item{dCb2}{numeric, stable isotope values of d13C for baseline 2}
  \item{dNb2}{numeric, stable isotope values of d15N for baseline 2}
  \item{deltaN}{numeric, trophic discrimination factor for nitrogen}
  \item{deltaC}{numeric, trophic discrimination factor for carbon}
}}
\usage{
data("Trout")
}
\description{
A dataset containing stable isotope values (d13C and d15N) for the invasive
trout *Oncorhynchus mykiss*
(www.fishbase.se/summary/oncorhynchus-mykiss.html), found in the
Chungara Lake at 4,500 meters above sea level, Chilean Altiplano.
}
\keyword{datasets}
