% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biPlot.R
\name{biPlot}
\alias{biPlot}
\title{Internal functions within tRophicPosition}
\usage{
biPlot(df = NULL, df2 = NULL, ylab = NULL, xlab = NULL, p = "p1",
  legend = legend, limits = limits, ...)
}
\arguments{
\item{...}{Multiple parameters.}
}
\value{
Multiple objects returned
}
\description{
These functions are intended to be used within the package, and not directly by the users
}
