% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biPlot.R
\name{biPlot}
\alias{biPlot}
\title{Function that creates biplot with marginal density functions}
\usage{
biPlot(df = NULL, df2 = NULL, ylab = NULL, xlab = NULL, p = "p1",
  legend = legend, limits = limits, ...)
}
\arguments{
\item{...}{}
}
\description{
This function is intended to be used within the package.
}
