% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNames.R
\name{checkNames}
\alias{checkNames}
\title{Internal function that check the names of a list.}
\usage{
checkNames(df = NULL, flag = NULL)
}
\arguments{
\item{df}{named list with the variables to be checked for.}

\item{flag}{integer (stating the number of baselines) or a list of names
(stating the names expected).}
}
\description{
Not intended to be used by the user.
}
\examples{

}

