% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credibilityIntervals.R
\name{credibilityIntervals}
\alias{credibilityIntervals}
\title{Title}
\usage{
credibilityIntervals(df, x = "species", plotAlpha = TRUE, y1 = "median",
  y1min = "lower", y1max = "upper", y2 = "alpha.median",
  y2min = "alpha.lower", y2max = "alpha.upper", xlab = "Bayesian models",
  ylab1 = "Posterior Trophic Position", ylab2 = "Posterior alpha")
}
\arguments{
\item{df}{}
}
\description{
Title
}
