% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generataTPData.R
\name{generateTPData}
\alias{generateTPData}
\title{A function to generate some data for trophic position calculation}
\usage{
generateTPData(n.obs = 25, std.dev = 0.1, n.baselines = 1, dNb1 = NULL,
  dNsc = NULL, deltaN = 3.4)
}
\arguments{
\item{deltaN}{}
}
\description{
A function to generate some data for trophic position calculation
}

