% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generataTPData.R
\name{generateTPData}
\alias{generateTPData}
\title{A function to generate random stable isotope data for trophic position
calculation}
\usage{
generateTPData(n.baselines = 2, n.obsB = 25, n.obsC = 25, dNb1 = NULL,
  sd.dNb1 = 0.5, dNb2 = NULL, sd.dNb2 = 0.5, dCb1 = NULL,
  sd.dCb1 = 0.5, dCb2 = NULL, sd.dCb2 = 0.5, dNc = NULL, sd.dNc = 0.5,
  dCc = NULL, sd.dCc = 0.5, n.obsDeltaN = 56, DeltaN = 3.4,
  sd.DeltaN = 0.98, n.obsDeltaC = 107, DeltaC = 0.39, sd.DeltaC = 1.3,
  consumer = NULL)
}
\arguments{
\item{n.baselines}{Number of baselines (could be 1 or 2), by default is 2.}

\item{n.obsB}{Number of observations for baselines. Default is 25.}

\item{n.obsC}{Number of observations for consumer. Default is 25.}

\item{dNb1}{mean value for delta Nitrogen of baseline 1. Default is a random
number between -5 and 5}

\item{dNc}{mean value for delta Nitrogen of secondary consumer. Default value
is dNb1 multiplied 2 times the trophic discrimination factor.}

\item{n.obsDeltaN}{Number of observations of deltaN (trophic discrimination
factor). Default is 56}

\item{DeltaN}{mean value for trophic discrimination factor. Default value is 3.4}

\item{n.obsDeltaC}{Number of observations of DeltaC (trophic discrimination
factor). Default is 107}

\item{DeltaC}{mean value for trophic discrimination factor. Default value is 0.39}

\item{consumer}{}
}
\value{
An isotopeData class object (named list) with dNb1, dNc and deltaN
randomly generated observations. If n.baselines = 2, then dCb1, dNb2, dCb2,
dCc and deltaC are also returned.
}
\description{
This function generates random stable isotope (d13C and d15N) data for using
basic functions of tRophicPosition package.
}
