% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsOneBaseline.R
\name{jagsOneBaseline}
\alias{jagsOneBaseline}
\title{Defines the jags model to fit the single baseline trophic position model}
\usage{
jagsOneBaseline(muB = NULL, sigmaB = NULL, muDeltaN = NULL,
  sigmaDeltaN = NULL, sigma = NULL, TP = NULL, lambda = NULL, ...)
}
\arguments{
\item{muB}{a distribution defining prior for mean (mu) of baseline.}

\item{sigmaB}{a distribution defining sigma (std dev) of baseline.}

\item{muDeltaN}{a distribution defining prior for the mean (mu) of
deltaN. deltaN stands for trophic enrichment factor of Nitrogen.}

\item{sigmaDeltaN}{a distribution defining sigma (std dev) of deltaN.}

\item{sigma}{a value defining sigma (std dev) of baseline.}

\item{TP}{a distribution defining prior of trophic position.}

\item{lambda}{an integer indicating the trophic position of the baseline.}
}
\value{
A jags model as a character string
}
\description{
Takes some parameters and returns a jags model object as a
character string for passing to \code{\link[rjags]{jags.model}}. Although
it is possible to use a number of predefined or customized
distributions (see
\href{https://sourceforge.net/projects/mcmc-jags/files/Manuals/}{JAGS documentation}),
 it is likelly that most of the time
you will be using a normal distribution. This is the default option (i.e.
when the function is called without arguments) and it is like this:
"mu ~ dnorm(0, 0.0001)". In this case, a prior of normally distributed mu is
defined, with a mean 0, and a standard deviation of 0.0001. This is a normal
distributed prior, although uninformative. You might want to change the mean
and/or the standard deviation according to your previously knowledge of the
system you are working on. As well as the prior for mu, JAGS uses "tau",
which is the precision. Precision is a deterministic function (instead of the
distributional "~"), and it is calculated as "tau <- power(sigma, -2)", thus
you have to define as well sigma, which stands for the standard deviation.
}
