% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsOneBaseline.R
\name{jagsOneBaseline}
\alias{jagsOneBaseline}
\title{Defines the jags model to fit the single baseline trophic position model}
\usage{
jagsOneBaseline(muBprior = NULL, sigmaBprior = NULL, muDeltaNprior = NULL,
  sigmaDeltaNprior = NULL, sigmaPrior = NULL, TPprior = NULL,
  lambda = NULL)
}
\arguments{
\item{muBprior}{a distribution defining prior for mu of Baseline.}

\item{sigmaBprior}{}

\item{lambda}{an integer indicating the trophic position of the baseline.}
}
\value{
A jags model as a character string
}
\description{
Takes some parameters and returns a jags model object as a
character string for passing to \code{\link[rjags.model]{rjags.model}}.
}

