% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsOneBaseline.R
\name{jagsOneBaseline}
\alias{jagsOneBaseline}
\title{Defines the jags model to fit the single baseline trohpic position model}
\usage{
jagsOneBaseline()
}
\value{
A jags model as a character string

 @export
}
\description{
Takes no input parameters and returns a jags model object as a
character string for passing to \code{\link[rjags.model]{rjags.model}}.
}

