% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsOneBaseline.R
\name{jagsOneBaseline}
\alias{jagsOneBaseline}
\title{Defines the jags model to fit the single baseline trophic position model}
\usage{
jagsOneBaseline(muB = NULL, sigmaB = NULL, muDeltaN = NULL,
  sigmaDeltaN = NULL, sigma = NULL, TP = NULL, lambda = NULL)
}
\arguments{
\item{muB}{a distribution defining prior for mean (mu) of baseline.}

\item{sigmaB}{a distribution defining sigma (std dev) of baseline.}

\item{muDeltaN}{a distribution defining prior for the mean (mu) of
deltaN. deltaN stands for trophic enrichment factor of Nitrogen.}

\item{sigmaDeltaN}{a distribution defining sigma (std dev) of deltaN.}

\item{sigma}{a value defining sigma (std dev) of baseline.}

\item{TP}{a distribution defining prior of trophic position.}

\item{lambda}{an integer indicating the trophic position of the baseline.}
}
\value{
A jags model as a character string
}

