% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsTwoBaselines.R
\name{jagsTwoBaselines}
\alias{jagsTwoBaselines}
\title{Defines the jags model to fit the two baseline trophic position model}
\usage{
jagsTwoBaselines(muCb1 = NULL, sigmaCb1 = NULL, muNb1 = NULL,
  sigmaNb1 = NULL, muCb2 = NULL, sigmaCb2 = NULL, muNb2 = NULL,
  sigmaNb2 = NULL, alpha = NULL, sigmaCsc = NULL, TP = NULL,
  sigmaNsc = NULL, muDeltaN = NULL, sigmaDeltaN = NULL, lambda = NULL)
}
\arguments{
\item{lambda}{}
}
\value{
A jags model as a character string
}
\description{
Takes some parameters and returns a jags model object as a
character string for passing to \code{\link[rjags.model]{rjags.model}}.
}

