% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiModelTP.R
\name{multiModelTP}
\alias{multiModelTP}
\title{Multiple models calculation of trophic position}
\usage{
multiModelTP(siData = siData, lambda = 2, n.chains = 2, n.adapt = 10000,
  n.iter = 10000, burnin = 10000, thin = 10, models = c("oneBaseline",
  "twoBaselines", "twoBaselinesFull"), print = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{burnin}{only used in parallel version}

\item{...}{}
}
\description{
Multiple models calculation of trophic position
}
