% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiModelTP.R
\name{multiModelTP}
\alias{multiModelTP}
\title{Multiple models calculation of trophic position}
\usage{
multiModelTP(siData = siData, lambda = 2, n.chains = 2, n.adapt = 20000,
  n.iter = 20000, burnin = 20000, thin = 10, models = c("oneBaseline",
  "twoBaselines", "twoBaselinesFull"), print = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Multiple models calculation of trophic position
}
