% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiModelTP.R
\name{multiModelTP}
\alias{multiModelTP}
\title{Title}
\usage{
multiModelTP(siData = siData, lambda = 2, print = TRUE, n.adapt = 20000,
  n.iter = 20000, n.chains = 2, models = c("oneBaseline", "twoBaselines",
  "twoBaselinesFull"), parallel = NULL, burnin = 20000)
}
\arguments{
\item{burnin}{}
}
\description{
Title
}
