% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSpeciesTP.R
\name{multiSpeciesTP}
\alias{multiSpeciesTP}
\title{Multiple species calculation of trophic position}
\usage{
multiSpeciesTP(siDataList = siDataList, lambda = 2, print = FALSE,
  n.adapt = 20000, n.iter = 20000, n.chains = 5, burnin = 20000,
  models = c("oneBaseline", "twoBaselines", "twoBaselinesFull"),
  parallel = NULL)
}
\arguments{
\item{burnin}{}
}
\description{
Multiple species calculation of trophic position
}
