% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSpeciesTP.R
\name{multiSpeciesTP}
\alias{multiSpeciesTP}
\title{Multiple species calculation of trophic position}
\usage{
multiSpeciesTP(siDataList = siDataList, lambda = 2, n.chains = 2,
  n.adapt = 10000, n.iter = 10000, burnin = 10000, thin = 10,
  model = "oneBaseline", print = FALSE, ...)
}
\arguments{
\item{burnin}{only used in parallel - check that}

\item{...}{}
}
\description{
Multiple species calculation of trophic position
}
