% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSpeciesTP.R
\name{multiSpeciesTP}
\alias{multiSpeciesTP}
\title{Multiple species calculation of trophic position}
\usage{
multiSpeciesTP(siDataList = siDataList, lambda = 2, n.chains = 2,
  n.adapt = 20000, n.iter = 20000, burnin = 20000, thin = 10,
  model = "oneBaseline", print = FALSE, parallel = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Multiple species calculation of trophic position
}
