% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseComparisons.R
\name{pairwiseComparisons}
\alias{pairwiseComparisons}
\title{Function to perform a pairwise comparison between several posterior distributions}
\usage{
pairwiseComparisons(df, test = "<=", print = FALSE)
}
\arguments{
\item{test}{}
}
\description{
Function to perform a pairwise comparison between several posterior distributions
}
\examples{

}
