% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametricTP.R
\name{parametricTP}
\alias{parametricTP}
\title{Parametric trophic position}
\usage{
parametricTP(siData, lambda = 2, print = TRUE)
}
\arguments{
\item{siData}{an isotopeData class object.}

\item{lambda}{numerical value representing trophic level of baseline(s).}

\item{print}{a logical value to indicate whether the output is printed or
not.}
}
\value{
a list with parametric trophic position calculated with a one
baseline model, a two baselines model and its alpha value, and a two
baselines full model and its alpha value.
}
\description{
Calculation of parametric trophic position partially based on Post (2002).
}
\examples{
consumer <- generateTPData()
parametricTP(consumer)
}
