% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametricTP.R
\name{parametricTP}
\alias{parametricTP}
\title{Parametric trophic position}
\usage{
parametricTP(siData, lambda = 2, print = TRUE)
}
\arguments{
\item{df}{}
}
\description{
Calculation of old school trophic position
}

