% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.isotopeData.R
\name{plot.isotopeData}
\alias{plot.isotopeData}
\title{Title}
\usage{
\method{plot}{isotopeData}(IsotopeData, consumer = "Consumer",
  b1 = "Pelagic baseline", b2 = "Benthic baseline", legend = c(0.85,
  0.85), density = "both")
}
\arguments{
\item{density}{}
}
\description{
Title
}
\examples{

}

