% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTP.R
\name{plotTP}
\alias{plotTP}
\title{Function to plot a trophic position distribution}
\usage{
plotTP(TPdist = NULL, ...)
}
\arguments{
\item{TPdist}{One posterior distribution (or a collection) of trophic position.
In case of wanting to plot two or more posterior distributions, needs to be
passed as a {\link[base]{data.frame}} object.}

\item{...}{additional arguments passed to this function.}
}
\value{
A new figure window
}
\description{
Wrapper of {\link[SIBER]{siberDensityPlot}}.
}
\examples{
species1 <- stats::rnorm(1000, 4, 0.1)
species2 <- stats::rnorm(1000, 3, 0.8)
plotTP(data.frame(species1, species2))


}
