% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reOrder.R
\name{reOrder}
\alias{reOrder}
\title{Internal function that re-order variables from a data frame.}
\usage{
reOrder(df = NULL)
}
\arguments{
\item{df}{data frame containing the data to be ordered.}
}
\value{

}
\description{
Not intended to be used by the user.
}
