% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenIsotopeData.R
\name{screenIsotopeData}
\alias{screenIsotopeData}
\title{Function to plot and screen Isotope Data with one or more baselines.}
\usage{
screenIsotopeData(IsotopeData = NULL, density = "both")
}
\arguments{
\item{IsotopeData}{a named list composed at least of 6 vectors, dNb1, dCb1,
dCsc, dNsc, deltaN and deltaC.
dNb1 and dCb1 stands for delta Nitrogen and delta Carbon of baseline 1, dNsc
and dCsc stands for delta N and delta C of secondary consumer, and deltaN
and deltaC stands for the trophic enrichment factor (capital letter delta)
of Nitrogen and Carbon.}
}
\value{
none
}
\description{
This function receives a named list of vectors, and plots a scatterplot with
 2 sources and a consumer.
}

