% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenIsotopeData.R
\name{screenIsotopeData}
\alias{screenIsotopeData}
\title{Function to plot and screen Isotope Data with one or two baselines.}
\usage{
screenIsotopeData(TPdata = NULL)
}
\arguments{
\item{TPdata}{a named list composed of two vectors, dNb1 and dNsc.
dNb1 stands for delta Nitrogen of baseline 1, while dNsc stands for
delta Nitrogen of secondary consumer.}
}
\description{
This function receives a named list of two vectors, and plot a basic histogram.
}

