% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenIsotopeData.R
\name{screenIsotopeData}
\alias{screenIsotopeData}
\title{Function to plot and screen isotope data with one or more baselines.}
\usage{
screenIsotopeData(IsotopeData = NULL, density = "both",
  consumer = "Consumer", b1 = "Pelagic baseline", b2 = "Benthic baseline",
  legend = c(0.85, 0.85), ...)
}
\arguments{
\item{density}{a character that might plot the density function. Accepted
characters are "both" in which case will plot the density function above and
to the right, "right", "above" or "none".}

\item{...}{}
}
\value{
none
}
\description{
This function receives a named list of vectors, and plots a scatterplot with
 2 sources and a consumer. Depending on the length of the named list, this
 function plots one (length = 6), or two (length = 8) baselines. Also the
 user can states whether he/she wants a density function ploted above, to the
 right, both or does not want it to be ploted.
}

