% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenIsotopeData.R
\name{screenIsotopeData}
\alias{screenIsotopeData}
\title{Function to plot and screen Isotope Data with one or two baselines.}
\usage{
screenIsotopeData(IsotopeData = NULL, histogram = T, density = "both")
}
\arguments{
\item{IsotopeData}{a named list composed of two vectors, dNb1 and dNsc.
dNb1 stands for delta Nitrogen of baseline 1, while dNsc stands for
delta Nitrogen of secondary consumer.}
}
\value{
none
}
\description{
This function receives a named list of vectors, and plot a basic
histogram if the list has 3 elements. If it receives a list of 7 elements,
calls another function to plot a scatterplot with 2 sources and a consumer.
}

