% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenIsotopeData.R
\name{screenIsotopeData}
\alias{screenIsotopeData}
\title{Function to plot and screen Isotope Data with one or more baselines.}
\usage{
screenIsotopeData(IsotopeData = NULL, density = "both")
}
\arguments{
\item{IsotopeData}{a named list composed at least of 6 vectors, dNb1, dCb1,
dCc, dNc, deltaN and deltaC.
dNb1 and dCb1 stands for delta Nitrogen and delta Carbon of baseline 1, dNc
and dCc stands for delta N and delta C of consumer, and deltaN
and deltaC stands for the trophic enrichment factor (capital letter delta)
of Nitrogen and Carbon.}

\item{density}{a character that might plot the density function. Accepted
characters are "both" in which case will plot the density function above and
to the right, "right", "above" or "none".}
}
\value{
none
}
\description{
This function receives a named list of vectors, and plots a scatterplot with
 2 sources and a consumer. Depending on the length of the named list, this
 function plots one (length = 6), or two (length = 8) baselines. Also the
 user can states whether he/she wants a density function ploted above, to the
 right, both or does not want it to be ploted.
}

