% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenIsotopeData1source.R
\name{screenIsotopeData1source}
\alias{screenIsotopeData1source}
\title{Internal function that plots the data}
\usage{
screenIsotopeData1source(isotopeData = NULL, type = "histogram")
}
\arguments{
\item{isotopeData}{a named list composed at least of 6 vectors, dNb1, dCb1,
dCc, dNc, deltaN and deltaC}

\item{type}{character that states if an "histogram" or a "density" will be
ploted.}
}
\description{
This function is intended to be used within the package. For details refer to
\code{\link{screenIsotopeData}}. Is planned to deprecate it in the near
future.
}
\examples{

}

