% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenIsotopeData3sources.R
\name{screenIsotopeData3sources}
\alias{screenIsotopeData3sources}
\title{Title}
\usage{
screenIsotopeData3sources(isotopeData = NULL, density = TRUE)
}
\arguments{
\item{density}{}
}
\description{
Title
}
\note{

}
\examples{

}

