% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenIsotopeDataMoreSources.R
\name{screenIsotopeDataMoreSources}
\alias{screenIsotopeDataMoreSources}
\title{Title}
\usage{
screenIsotopeDataMoreSources(isotopeData = NULL, density = "both",
  baselines = 2)
}
\arguments{
\item{density}{}
}
\description{
Title
}
\note{

}
\examples{

}

