% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateTEF.R
\name{simulateTEF}
\alias{simulateTEF}
\title{Simulate trophic enrichment factors}
\usage{
simulateTEF(nN = 56, meanN = NULL, sdN = 0.98, nC = 107, meanC = NULL,
  sdC = 1.3)
}
\arguments{
\item{nN}{number of observations for deltaN}

\item{meanN}{mean for deltaN}

\item{sdN}{standard deviation for deltaN}

\item{nC}{number of observations for deltaC}

\item{meanC}{mean for deltaC}

\item{sdC}{standard deviation for deltaC}
}
\description{
\code{simulateTEF} return trophic enrichment factors (TEF), given a number of
observations, a mean and a standard deviation for DeltaN and/or DeltaC.
}
\examples{

}
