% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanOneBaselineCSIA.R
\name{stanOneBaselineCSIA}
\alias{stanOneBaselineCSIA}
\title{Title}
\usage{
stanOneBaselineCSIA(
  aminoacid = NULL,
  Spp = NULL,
  TDF = NULL,
  TDF_sd = 1.2,
  nBeta = 32,
  Beta = 3.4,
  Beta_sd = 0.68,
  iter = 10000,
  adapt_delta = 0.9
)
}
\arguments{
\item{adapt_delta}{}
}
\value{
value
}
\description{
Title
}
