% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.isotopeData.R
\name{summary.isotopeData}
\alias{summary.isotopeData}
\title{Summary for stable isotope data}
\usage{
\method{summary}{isotopeData}(siData, print = TRUE, round_dec = 1)
}
\arguments{
\item{siData}{}
}
\description{
Summary for stable isotope data
}
